harness: split null.h in null.h and null.c
